+++ CHANGELOG.md
@@ -4,6 +4,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 
 
 ### Unreleased
+* Feature - Service marketplace UI Updates
 * Technical User
    * Removed multi overlay layering
    * Update the style of the response